Update network_walk to add a function to get component_ids and change the .tsv separator to "," instead to be able to merge them through command line
Update network_walk to add a function to get component_ids and change the .tsv separator to "," instead to be able to merge them through command line. Furthermore, the string fields are quoted "" to avoid misformats. There is a comment for Ricardo to check the line 194. There is a mistake that was not supposed to be there and it was not tested because usually the code is run by task id with a network